requirements.txt
# ============================================================
# Personal AI Memory System — Dependencies
# ============================================================

# WordPress REST API & HTTP
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# HTML parsing (strip HTML from WordPress content)
beautifulsoup4>=4.12.0

# Data handling (CSV ingestion)
pandas>=2.1.0

# OpenAI embeddings
openai>=1.12.0

# ChromaDB vector database
chromadb>=0.4.22

# Neo4j graph database driver
neo4j>=5.17.0

# LangChain framework + Groq LLM provider
langchain>=0.1.0
langchain-groq>=0.0.3
langchain-core>=0.1.0

# Dashboard UI (Backend)
Flask>=3.0.0
flask-cors>=4.0.0
flasgger>=0.9.7.1

# Interactive graph visualization
pyvis>=0.3.2

# Scheduling (optional daily runs)
schedule>=1.2.0
